Gradle: Gather common Kotlin DSL extension functions in the buildSrc project
