+++ .gitignore
@@ -1,4 +1,4 @@
-/.gradle/
+.gradle/
 .idea/
 build/
 out/
