+++ spdx-utils/build.gradle.kts
@@ -1,12 +1,9 @@
 import at.bxm.gradleplugins.svntools.tasks.SvnExport
 
+import com.here.ort.gradle.*
+
 import groovy.json.JsonSlurper
 
-import org.gradle.plugins.ide.idea.model.IdeaModel
-import org.gradle.plugins.ide.idea.model.IdeaProject
-
-import org.jetbrains.gradle.ext.ProjectSettings
-import org.jetbrains.gradle.ext.TaskTriggersConfig
 import org.jetbrains.kotlin.gradle.dsl.KotlinCompile
 
 import java.io.FileFilter
@@ -34,17 +31,6 @@ val generateGrammarSource by tasks.existing(AntlrTask::class) {
 tasks.withType(KotlinCompile::class) {
     dependsOn(generateGrammarSource)
 }
-
-// The following extension functions add the missing Kotlin DSL syntactic sugar for configuring the idea-ext plugin.
-
-fun Project.idea(block: IdeaModel.() -> Unit) =
-    (this as ExtensionAware).extensions.configure("idea", block)
-
-fun IdeaProject.settings(block: ProjectSettings.() -> Unit) =
-    (this@settings as ExtensionAware).extensions.configure(block)
-
-fun ProjectSettings.taskTriggers(block: TaskTriggersConfig.() -> Unit) =
-    (this@taskTriggers as ExtensionAware).extensions.configure("taskTriggers", block)
 
 rootProject.idea {
     project {
